+++ engine/include/sprite/spriteframe.hpp
@@ -7,10 +7,10 @@ namespace engine::sprite
     struct SpriteFrame
     {
         unsigned int id;
-        double duration;
+        float duration;
         engine::Rectangle source_rect;
 
-        SpriteFrame(unsigned int id, double duration, const engine::Rectangle &source_rect)
+        SpriteFrame(unsigned int id, float duration, const engine::Rectangle &source_rect)
             : id(id), duration(duration), source_rect(source_rect)
         {
         }
